+++ inst/doc/readme.R
@@ -517,7 +517,8 @@ fn
 #      "x" = list(),
 #      "y" = list(),
 #      "z" = list(10)
-#    )
+#    ),
+#    identation = 2
 #  )
 #  
 #  #